Miscellaneous fixes
I made several tweaks while running on the 19 river set from Durand et al (2016). The main change is reading in the time window for the inversion as actual times, matched up to the times in the observation input... before was treating as indices.

Durand, M., Gleason, C., Garambois, P., Bjerklie, D., Smith, L., Roux, H., Rodriguez, E., Bates, P., Pavelsky, T., Monnier, J., Chen, X., Baldassarre, G., Fiset, J., Flipo, N., Frasson, R., Fulton, J., Goutal, N., Hossain, F., Humphries, E., Minear, J., Mukolwe, M., Neal, J., Ricci, S., Sanders, B., Schumann, G., Schubert, J., Vilmin, L. (2016). An intercomparison of remote sensing river discharge estimation algorithms from measurements of river height, width, and slope Water Resources Research  52(6), 4527-4549. https://dx.doi.org/10.1002/2015wr018434

+++ ReadParams.py
@@ -16,7 +16,7 @@ def ReadParams(fname,D):
     R.Seed=eval(infile[5])
     
     Exp=Experiment()
-    buf=infile[7]; buf=buf.split(); Exp.tUse=array(buf,int)
+    buf=infile[7]; buf=buf.split(); Exp.tUse=array(buf,float)
     
     P=Prior(D)
     P.meanQbar=eval(infile[9])
